Refactor mesh module into package structure
