+++ data/mesh/__main__.py
@@ -1,0 +1,7 @@
+"""Command-line entry point for the PotatoMesh daemon."""
+
+from .daemon import main
+
+
+if __name__ == "__main__":
+    main()
